backend/requirements.txt
alembic
annotated-types
anyio
argon2-cffi
argon2-cffi-bindings
arrow
asttokens
async-generator
async-lru
attrs
Babel
beautifulsoup4
bleach
blinker
Brotli
cachelib
celery
certifi
certipy
cffi
charset-normalizer
click
comm
contourpy
cryptography
cycler
dash
dash-bootstrap-components
dash-core-components
dash-draggable
dash-extensions
dash-html-components
dash-table
dataclass-wizard
debugpy
decorator
defusedxml
distro
django
djangorestframework
django-cors-headers
django-environ
djangorestframework-simplejwt
psycopg2-binary
EditorConfig
executing
fastjsonschema
fonttools
fqdn
greenlet
gunicorn
h11
httpcore
httpx
idna
importlib_metadata
ipykernel
ipython
ipywidgets
isoduration
itsdangerous
jedi
Jinja2
jiter
joblib
jsbeautifier
json5
jsonpointer
jsonschema
jsonschema-specifications
jupyter-console
jupyter-events
jupyter-lsp
jupyter-telemetry
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterhub
jupyterlab
jupyterlab-pygments
jupyterlab-widgets
jupyterlab_server
kiwisolver
Mako
MarkupSafe
matplotlib
matplotlib-inline
mistune
mkl-service
more-itertools
nbclient
nbconvert
nbformat
nest-asyncio
networkx
notebook
notebook_shim
numpy
oauthlib
openai
opencv-python
overrides
packaging
pam
pamela
pandas
pandocfilters
parso
pexpect
pillow
platformdirs
plotly
ply
prometheus-client
prompt-toolkit
psutil
ptyprocess
pure-eval
pycparser
pycurl
pydantic
pydantic_core
Pygments
PyJWT
pyOpenSSL
pyparsing
PyQt5
PyQt5-sip
PySocks
pytesseract
python-dateutil
python-dotenv
python-json-logger
python-pam
pytz
PyYAML
pyzmq
qtconsole
QtPy
redis
referencing
requests
retrying
rfc3339-validator
rfc3986-validator
rpds-py
ruamel.yaml
ruff
scikit-learn
scipy
seaborn
Send2Trash
setuptools
sip
six
sniffio
soupsieve
SQLAlchemy
stack-data
tenacity
terminado
threadpoolctl
tinycss2
tornado
tqdm
traitlets
types-python-dateutil
typing_extensions
tzdata
uri-template
urllib3
wcwidth
webcolors
webencodings
websocket-client
Werkzeug
wheel
widgetsnbextension
zipp